disallow signing up / canceling signups for past appointments
Test plan:
 0. Enable New Scheduler and create an appointment group with
    a slot that will end in the next few minutes (it doesn't let
    you create one that's already ended)
 1. Hurry and sign up a student berfore the end time.
 2. Wait for the appointment time to pass.
 3. Ensure the student does not see an "Un-reserve" link
    when clicking their reservation.
 4. Ensure a different student (who is not signed up for
    the slot) doesn't see it in Find Appointment mode.
 5. ensure the Calendar API endpoint "Reserve a time slot"
    returns 403 Forbidden if a student tries to reserve
    a slot that ended in the past
 6. ensure the Calendar API endpoint "Delete a calendar event"
    returns 403 Forbidden if a student tries to delete
    a reservation for a time slot that ended in the past
    (use the student's reservation event here, not the
     time slot)
 7. ensure the Calendar API endpoint "Delete a calendar event"
    works when a teacher tries to delete a reservation
    for a student that ended in the past

closes CNVS-35537

Reviewed-on: https://gerrit.instructure.com/110582
Tested-by: Jenkins
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Chris Ward <cward@instructure.com>

+++ app/controllers/calendar_events_api_controller.rb
@@ -514,7 +514,7 @@ class CalendarEventsApiController < ApplicationController
   #        -H "Authorization: Bearer <token>"
   def reserve
     get_event
-    if authorized_action(@event, @current_user, :reserve)
+    if authorized_action(@event, @current_user, :reserve) && check_for_past_signup(@event)
       begin
         participant_id = Shard.relative_id_for(params[:participant_id], Shard.current, Shard.current) if params[:participant_id]
         if participant_id && @event.appointment_group.grants_right?(@current_user, session, :manage)
@@ -644,7 +644,7 @@ class CalendarEventsApiController < ApplicationController
   #        -H "Authorization: Bearer <token>"
   def destroy
     get_event
-    if authorized_action(@event, @current_user, :delete)
+    if authorized_action(@event, @current_user, :delete) && check_for_past_signup(@event.parent_event)
       @event.updating_user = @current_user
       @event.cancel_reason = params[:cancel_reason]
       if @event.destroy
@@ -1292,5 +1292,15 @@ class CalendarEventsApiController < ApplicationController
   def calendar_event_params
     params.require(:calendar_event).
       permit(CalendarEvent.permitted_attributes + [:child_event_data => strong_anything])
+  end
+
+  def check_for_past_signup(event)
+    if event && event.end_at < Time.now.utc && event.context.is_a?(AppointmentGroup)
+      unless event.context.grants_right?(@current_user, :manage)
+        render :json => { :message => 'Cannot create or change reservation for past appointment' }, :status => :forbidden
+        return false
+      end
+    end
+    true
   end
 end
